Move common logic to utils
