+++ packages/eslint-plugin-kolibri/lib/rules/vue-no-unused-properties.js
@@ -7,29 +7,16 @@
 const remove = require('lodash/remove');
 const eslintPluginVueUtils = require('eslint-plugin-vue/lib/utils');
 
+const utils = require('../utils');
+
 const GROUP_PROPERTY = 'props';
 const GROUP_DATA = 'data';
 const GROUP_COMPUTED_PROPERTY = 'computed';
-const GROUP_WATCHER = 'watch';
 
 const PROPERTY_LABEL = {
   [GROUP_PROPERTY]: 'property',
   [GROUP_DATA]: 'data',
   [GROUP_COMPUTED_PROPERTY]: 'computed property',
-};
-
-const getReferencesNames = references => {
-  if (!references || !references.length) {
-    return [];
-  }
-
-  return references.map(reference => {
-    if (!reference.id || !reference.id.name) {
-      return;
-    }
-
-    return reference.id.name;
-  });
 };
 
 const reportUnusedProperties = (context, properties) => {
@@ -47,18 +34,12 @@ const reportUnusedProperties = (context, properties) => {
 
 const create = context => {
   let hasTemplate;
-  let rootTemplateEnd;
   let unusedProperties = [];
   let thisExpressionsVariablesNames = [];
 
   const initialize = {
     Program(node) {
-      if (context.parserServices.getTemplateBodyTokenStore == null) {
-        context.report({
-          loc: { line: 1, column: 0 },
-          message:
-            'Use the latest vue-eslint-parser. See also https://vuejs.github.io/eslint-plugin-vue/user-guide/#what-is-the-use-the-latest-vue-eslint-parser-error.',
-        });
+      if (!utils.checkVueEslintParser(context)) {
         return;
       }
 
@@ -68,13 +49,9 @@ const create = context => {
 
   const scriptVisitor = Object.assign(
     {},
-    {
-      'MemberExpression[object.type="ThisExpression"][property.type="Identifier"][property.name]'(
-        node
-      ) {
-        thisExpressionsVariablesNames.push(node.property.name);
-      },
-    },
+    utils.executeOnThisExpressionProperty(property => {
+      thisExpressionsVariablesNames.push(property.name);
+    }),
     eslintPluginVueUtils.executeOnVue(context, obj => {
       unusedProperties = Array.from(
         eslintPluginVueUtils.iterateProperties(
@@ -83,10 +60,7 @@ const create = context => {
         )
       );
 
-      const watchers = Array.from(
-        eslintPluginVueUtils.iterateProperties(obj, new Set([GROUP_WATCHER]))
-      );
-      const watchersNames = watchers.map(watcher => watcher.name);
+      const watchersNames = utils.getWatchersNames(obj);
 
       remove(unusedProperties, property => {
         return (
@@ -101,33 +75,23 @@ const create = context => {
     })
   );
 
-  const templateVisitor = {
-    'VExpressionContainer[expression!=null][references]'(node) {
-      const referencesNames = getReferencesNames(node.references);
+  const templateVisitor = Object.assign(
+    {},
+    {
+      'VExpressionContainer[expression!=null][references]'(node) {
+        const referencesNames = utils.getReferencesNames(node.references);
 
-      remove(unusedProperties, property => {
-        return referencesNames.includes(property.name);
-      });
+        remove(unusedProperties, property => {
+          return referencesNames.includes(property.name);
+        });
+      },
     },
-    // save root template end location - just a helper to be used
-    // for a decision if a parser reached the end of the root template
-    "VElement[name='template']"(node) {
-      if (rootTemplateEnd) {
-        return;
-      }
-
-      rootTemplateEnd = node.loc.end;
-    },
-    "VElement[name='template']:exit"(node) {
-      if (node.loc.end !== rootTemplateEnd) {
-        return;
-      }
-
+    utils.executeOnRootTemplateEnd(() => {
       if (unusedProperties.length) {
         reportUnusedProperties(context, unusedProperties);
       }
-    },
-  };
+    })
+  );
 
   return Object.assign(
     {},
