+++ packages/eslint-plugin-kolibri/lib/rules/vue-no-unused-vuex-properties.js
@@ -7,23 +7,10 @@
 const remove = require('lodash/remove');
 const eslintPluginVueUtils = require('eslint-plugin-vue/lib/utils');
 
-const GROUP_WATCHER = 'watch';
+const utils = require('../utils');
+
 const GETTER = 'getter';
 const STATE = 'state';
-
-const getReferencesNames = references => {
-  if (!references || !references.length) {
-    return [];
-  }
-
-  return references.map(reference => {
-    if (!reference.id || !reference.id.name) {
-      return;
-    }
-
-    return reference.id.name;
-  });
-};
 
 const reportUnusedVuexProperties = (context, properties) => {
   if (!properties || !properties.length) {
@@ -40,18 +27,12 @@ const reportUnusedVuexProperties = (context, properties) => {
 
 const create = context => {
   let hasTemplate;
-  let rootTemplateEnd;
   let unusedVuexProperties = [];
   let thisExpressionsVariablesNames = [];
 
   const initialize = {
     Program(node) {
-      if (context.parserServices.getTemplateBodyTokenStore == null) {
-        context.report({
-          loc: { line: 1, column: 0 },
-          message:
-            'Use the latest vue-eslint-parser. See also https://vuejs.github.io/eslint-plugin-vue/user-guide/#what-is-the-use-the-latest-vue-eslint-parser-error.',
-        });
+      if (!utils.checkVueEslintParser(context)) {
         return;
       }
 
@@ -60,13 +41,7 @@ const create = context => {
   };
 
   const scriptVisitor = Object.assign(
-    {
-      'MemberExpression[object.type="ThisExpression"][property.type="Identifier"][property.name]'(
-        node
-      ) {
-        thisExpressionsVariablesNames.push(node.property.name);
-      },
-    },
+    {},
     /*
       computed: mapState({
         count: state => state.todosCount
@@ -145,11 +120,11 @@ const create = context => {
         });
       },
     },
+    utils.executeOnThisExpressionProperty(property => {
+      thisExpressionsVariablesNames.push(property.name);
+    }),
     eslintPluginVueUtils.executeOnVue(context, obj => {
-      const watchers = Array.from(
-        eslintPluginVueUtils.iterateProperties(obj, new Set([GROUP_WATCHER]))
-      );
-      const watchersNames = watchers.map(watcher => watcher.name);
+      const watchersNames = utils.getWatchersNames(obj);
 
       remove(unusedVuexProperties, property => {
         return (
@@ -164,33 +139,23 @@ const create = context => {
     })
   );
 
-  const templateVisitor = {
-    'VExpressionContainer[expression!=null][references]'(node) {
-      const referencesNames = getReferencesNames(node.references);
+  const templateVisitor = Object.assign(
+    {},
+    {
+      'VExpressionContainer[expression!=null][references]'(node) {
+        const referencesNames = utils.getReferencesNames(node.references);
 
-      remove(unusedVuexProperties, property => {
-        return referencesNames.includes(property.name);
-      });
+        remove(unusedVuexProperties, property => {
+          return referencesNames.includes(property.name);
+        });
+      },
     },
-    // save root template end location - just a helper to be used
-    // for a decision if a parser reached the end of the root template
-    "VElement[name='template']"(node) {
-      if (rootTemplateEnd) {
-        return;
-      }
-
-      rootTemplateEnd = node.loc.end;
-    },
-    "VElement[name='template']:exit"(node) {
-      if (node.loc.end !== rootTemplateEnd) {
-        return;
-      }
-
+    utils.executeOnRootTemplateEnd(() => {
       if (unusedVuexProperties.length) {
         reportUnusedVuexProperties(context, unusedVuexProperties);
       }
-    },
-  };
+    })
+  );
 
   return Object.assign(
     {},
